BULGARIAN BIBLE Copyright © 1995-2005 by bibliata.com
This digital version is based on the 1940 edition of the Bulgarian Bible.
The original text has been revised according to the current Bulgarian alphabet. The applied transliteration method is patented and may not be used without an express written permission from the patentee. The current version contains exactly: 66 books, 1,189 chapters, 31,101 verses, 679,466 words and 3,027,900 characters. It may be reproduced electronically only in its entirety, except in cases of direct quotation accompanied with complete bibliographical note. The text may be digitally formatted only if this does not modify the order and meaning of words, phrases and sentences. While the order of the books may be rearranged according to the Western canon, the original chapter and verse numbering must be preserved. Titles may be translated.

The text contains the following technical markings:
<> Translators note representing text not contained in MSS
[] Text not found in all known MSS
{} Meaning of the preceding word or phrase
() Grammatical function (in parentheses)